chore(v2-modal): Move modal to App level to prevent it displaying multiple times within site
