+++ src/components/PageHeader/index.tsx
@@ -4,7 +4,6 @@ import { Heading, IconButton, Text, Flex, useModal, TuneIcon, HistoryIcon } from
 import useI18n from 'hooks/useI18n'
 import SettingsModal from './SettingsModal'
 import RecentTransactionsModal from './RecentTransactionsModal'
-import UseV2ExchangeModal from '../UseV2ExchangeModal'
 
 interface PageHeaderProps {
   title: ReactNode
@@ -23,20 +22,8 @@ const Details = styled.div`
 
 const PageHeader = ({ title, description, children }: PageHeaderProps) => {
   const TranslateString = useI18n()
-  const [hasSeenModal, setHasSeenModal] = useState(false)
   const [onPresentSettings] = useModal(<SettingsModal translateString={TranslateString} />)
   const [onPresentRecentTransactions] = useModal(<RecentTransactionsModal translateString={TranslateString} />)
-  const [onPresentUseV2ExchangeModal] = useModal(<UseV2ExchangeModal />)
-
-  useEffect(() => {
-    const showModal = () => {
-      onPresentUseV2ExchangeModal()
-      setHasSeenModal(true)
-    }
-    if (!hasSeenModal) {
-      showModal()
-    }
-  }, [onPresentUseV2ExchangeModal, hasSeenModal])
 
   return (
     <StyledPageHeader>
