+++ src/pages/App.tsx
@@ -2,7 +2,7 @@ import React, { Suspense, useEffect, useState } from 'react'
 import { HashRouter, Route, Switch } from 'react-router-dom'
 import styled from 'styled-components'
 import { Credentials, StringTranslations } from '@crowdin/crowdin-api-client'
-import { LangType } from '@pancakeswap-libs/uikit'
+import { LangType, useModal } from '@pancakeswap-libs/uikit'
 import VersionBar from 'components/VersionBar'
 import Popups from '../components/Popups'
 import Web3ReactManager from '../components/Web3ReactManager'
@@ -18,6 +18,7 @@ import { RedirectPathToSwapOnly } from './Swap/redirects'
 import { EN, allLanguages } from '../constants/localisation/languageCodes'
 import { LanguageContext } from '../hooks/LanguageContext'
 import { TranslationsContext } from '../hooks/TranslationsContext'
+import UseV2ExchangeModal from '../components/UseV2ExchangeModal'
 
 import Menu from '../components/Menu'
 import useGetDocumentTitlePrice from '../hooks/useGetDocumentTitlePrice'
@@ -48,12 +49,24 @@ export default function App() {
   const apiKey = `${process.env.REACT_APP_CROWDIN_APIKEY}`
   const projectId = parseInt(`${process.env.REACT_APP_CROWDIN_PROJECTID}`)
   const fileId = 6
-
   const credentials: Credentials = {
     token: apiKey,
   }
 
   const stringTranslationsApi = new StringTranslations(credentials)
+
+  const [hasSeenModal, setHasSeenModal] = useState(false)
+  const [onPresentUseV2ExchangeModal] = useModal(<UseV2ExchangeModal />)
+
+  useEffect(() => {
+    const showModal = () => {
+      onPresentUseV2ExchangeModal()
+      setHasSeenModal(true)
+    }
+    if (!hasSeenModal) {
+      showModal()
+    }
+  }, [onPresentUseV2ExchangeModal, hasSeenModal])
 
   const getStoredLang = (storedLangCode: string) => {
     return allLanguages.filter((language) => {
